pyproject.toml
[project]
name = "outside-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.1",
    "aiosmtplib>=4.0.2",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "orjson>=3.11.2",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
dev = "src.scripts.server:run_server"
makemigrations = "src.scripts.aerich:makemigrations"
migrate = "src.scripts.aerich:migrate"
seed = "src.scripts.seed:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src",

]

[tool.aerich]
tortoise_orm = "src.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


